requirements.txt
openpyxl~=3.1.2
python-dotenv~=1.0.1
pydantic_settings
psycopg2-binary